+++ components/AvalancheForecastZoneMap.tsx
@@ -27,7 +27,8 @@ import {TravelAdvice} from 'components/helpers/travelAdvice';
 import {BodySmSemibold, Caption1, Caption1Black, Title3Black} from 'components/text';
 import {useAvalancheCenterMetadata} from 'hooks/useAvalancheCenterMetadata';
 import {useMapLayer} from 'hooks/useMapLayer';
-import {useMapLayerAvalancheForecasts, useMapLayerAvalancheWarnings} from 'hooks/useMapLayerForecasts';
+import {useMapLayerAvalancheForecasts} from 'hooks/useMapLayerAvalancheForecasts';
+import {useMapLayerAvalancheWarnings} from 'hooks/useMapLayerAvalancheWarnings';
 import {SafeAreaView} from 'react-native-safe-area-context';
 import {HomeStackNavigationProps} from 'routes';
 import {AvalancheCenterID, DangerLevel} from 'types/nationalAvalancheCenter';
