+++ hooks/useMapLayerAvalancheForecasts.ts
@@ -1,7 +1,6 @@
 import {useQueries, useQueryClient} from '@tanstack/react-query';
 import {ClientContext, ClientProps} from 'clientContext';
 import AvalancheForecastQuery from 'hooks/useAvalancheForecast';
-import AvalancheWarningQuery from 'hooks/useAvalancheWarning';
 import React from 'react';
 import {AvalancheCenter, AvalancheCenterID, MapLayer} from 'types/nationalAvalancheCenter';
 import {RequestedTime} from 'utils/date';
@@ -19,22 +18,6 @@ export const useMapLayerAvalancheForecasts = (center_id: AvalancheCenterID, requ
             queryKey: AvalancheForecastQuery.queryKey(nationalAvalancheCenterHost, center_id, feature.id, requestedTime, expiryTimeZone, expiryTimeHours),
             queryFn: async () => AvalancheForecastQuery.fetch(queryClient, nationalAvalancheCenterHost, center_id, feature.id, requestedTime, expiryTimeZone, expiryTimeHours),
             enabled: !!expiryTimeHours,
-            cacheTime: 24 * 60 * 60 * 1000, // hold on to this cached data for a day (in milliseconds)
-          };
-        })
-      : [],
-  });
-};
-
-export const useMapLayerAvalancheWarnings = (center_id: AvalancheCenterID, requestedTime: RequestedTime, mapLayer: MapLayer) => {
-  const {nationalAvalancheCenterHost} = React.useContext<ClientProps>(ClientContext);
-
-  return useQueries({
-    queries: mapLayer
-      ? mapLayer.features.map(feature => {
-          return {
-            queryKey: AvalancheWarningQuery.queryKey(nationalAvalancheCenterHost, center_id, feature.id, requestedTime),
-            queryFn: async () => AvalancheWarningQuery.fetch(nationalAvalancheCenterHost, center_id, feature.id, requestedTime),
             cacheTime: 24 * 60 * 60 * 1000, // hold on to this cached data for a day (in milliseconds)
           };
         })
